Refactor: Cria pasta para o componente form
