+++ src/components/form/form.module.css
@@ -1,0 +1,20 @@
+.container__form {
+  position: absolute;
+  width: 100%;
+  height: 80%;
+  left: 0px;
+  top: 50px;
+  display: flex;
+  align-items: center;
+  justify-content: center;
+}
+
+form {
+  display: flex;
+  flex-direction: column;
+  justify-content: space-evenly;
+  align-items: center;
+  width: 80%;
+  height: 80%;
+  background: #ECECEC;
+}
