+++ src/pages/Home.jsx
@@ -1,5 +1,5 @@
 import React from 'react';
-import Form from '../components/Form';
+import Form from '../components/form/Form';
 
 function Home() {
   return (
